글 수정, 삭제를 작성자 본인만 가능하도록 js 변경

+++ src/main/resources/static/js/app/index.js
@@ -34,40 +34,54 @@ var main = {
         });
     },
     update : function () {
+
         var data = {
             title: $('#title').val(),
             content: $('#content').val()
         };
-
+        var author = $('#author').val();
+        var username = $('#loginName').val();
         var id = $('#id').val();
 
-        $.ajax({
-            type: 'PUT',
-            url: '/api/v1/posts/'+id,
-            dataType: 'json',
-            contentType:'application/json; charset=utf-8',
-            data: JSON.stringify(data)
-        }).done(function() {
-            alert('글이 수정되었습니다.');
+        if(username == author) {
+            $.ajax({
+                type: 'PUT',
+                url: '/api/v1/posts/'+id,
+                dataType: 'json',
+                contentType:'application/json; charset=utf-8',
+                data: JSON.stringify(data)
+            }).done(function() {
+                alert('글이 수정되었습니다.');
+                window.location.href = '/';
+            }).fail(function (error) {
+                alert(JSON.stringify(error));
+            });
+        } else {
+            alert('본인만 수정할 수 있습니다.');
             window.location.href = '/';
-        }).fail(function (error) {
-            alert(JSON.stringify(error));
-        });
+        }
     },
     delete : function () {
         var id = $('#id').val();
-
-        $.ajax({
-            type: 'DELETE',
-            url: '/api/v1/posts/'+id,
-            dataType: 'json',
-            contentType:'application/json; charset=utf-8'
-        }).done(function() {
-            alert('글이 삭제되었습니다.');
+        var author = $('#author').val();
+        var username = $('#loginName').val();
+        if(username == author) {
+            $.ajax({
+                type: 'DELETE',
+                url: '/api/v1/posts/'+id,
+                dataType: 'json',
+                contentType:'application/json; charset=utf-8'
+            }).done(function() {
+                alert('글이 삭제되었습니다.');
+                window.location.href = '/';
+            }).fail(function (error) {
+                alert(JSON.stringify(error));
+            });
+        } else {
+            alert('본인만 삭제할 수 있습니다.');
             window.location.href = '/';
-        }).fail(function (error) {
-            alert(JSON.stringify(error));
-        });
+        }
+
     }
 
 };
